Добавлен просмотр всех сотрудников

+++ app/Http/Controllers/RoleController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\ApiException;
 use App\Http\Requests\CreateRoleRequest;
 use App\Http\Requests\UpdateRoleRequest;
 use App\Models\Role;
@@ -27,6 +28,9 @@ class RoleController extends Controller
     }
     public function deleteRole(int $id) {
         $role = Role::where('id', $id)->first();
+        if(!$role) {
+            throw new ApiException(404, 'Не найдено');
+        }
         $role->delete();
         return response()->json(['message' => 'Роль '.$id. ' удалена'])->setStatusCode(200);
     }
